+++ src/red/team/boss/list_boss/android/Android19.java
@@ -16,17 +16,16 @@ import red.services.Service;
 import red.services.TaskService;
 import red.utils.Util;
 
-
 public class Android19 extends Boss {
 
     public Android19() throws Exception {
         super(BossID.ANDROID_19, BossesData.ANDROID_19);
     }
 
-     @Override
+    @Override
     public void reward(Player plKill) {
-        int[] itemDos = new int[]{457,555,556,557,558,559,560,561,562,563,564,565,566,567};
-        int[] NRs = new int[]{15,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16};
+        int[] itemDos = new int[]{457, 555, 556, 557, 558, 559, 560, 561, 562, 563, 564, 565, 566, 567};
+        int[] NRs = new int[]{15, 16, 16, 16, 16, 16, 16, 16, 16, 16, 16, 16, 16, 16, 16, 16};
         int randomDo = new Random().nextInt(itemDos.length);
         int randomNR = new Random().nextInt(NRs.length);
         if (Util.isTrue(10, 100)) {
@@ -37,7 +36,7 @@ public class Android19 extends Boss {
 //                Service.gI().dropItemMap(this.zone, Util.ratiItem(zone, 2000+plKill.gender, 1, this.location.x, this.location.y, plKill.id));
 //            }
             if (Util.isTrue(1, 3)) {
-                Service.gI().dropItemMap(this.zone, Util.ratiItem(zone, 457, 1, this.location.x+2, this.location.y, plKill.id));
+                Service.gI().dropItemMap(this.zone, Util.ratiItem(zone, 457, 1, this.location.x + 2, this.location.y, plKill.id));
             }
             Service.gI().dropItemMap(this.zone, Util.ratiItem(zone, itemDos[randomDo], 1, this.location.x, this.location.y, plKill.id));
         } else {
@@ -50,7 +49,7 @@ public class Android19 extends Boss {
     @Override
     public void active() {
         super.active(); //To change body of generated methods, choose Tools | Templates.
-        if(Util.canDoWithTime(st,1800000)){
+        if (Util.canDoWithTime(st, 1800000)) {
             this.changeStatus(BossStatus.LEAVE_MAP);
         }
     }
@@ -58,12 +57,13 @@ public class Android19 extends Boss {
     @Override
     public void joinMap() {
         super.joinMap(); //To change body of generated methods, choose Tools | Templates.
-        st= System.currentTimeMillis();
+        st = System.currentTimeMillis();
     }
     private long st;
-   @Override
-    public int injured(Player plAtt, int damage, boolean piercing, boolean isMobAttack) {
- this.checkAnThan(plAtt);
+
+    @Override
+    public long injured(Player plAtt, long damage, boolean piercing, boolean isMobAttack) {
+        this.checkAnThan(plAtt);
         if (plAtt != null) {
             switch (plAtt.playerSkill.skillSelect.template.id) {
                 case Skill.KAMEJOKO:
@@ -80,12 +80,12 @@ public class Android19 extends Boss {
         return super.injured(plAtt, damage, piercing, isMobAttack);
     }
 
-   @Override
-   public void wakeupAnotherBossWhenDisappear() {
-       if (this.parentBoss != null) {
-           this.parentBoss.changeToTypePK();
-       }
-   }
+    @Override
+    public void wakeupAnotherBossWhenDisappear() {
+        if (this.parentBoss != null) {
+            this.parentBoss.changeToTypePK();
+        }
+    }
 
 }
