+++ src/red/team/boss/list_boss/conuong/Lytieunuong.java
@@ -62,7 +62,7 @@ public class Lytieunuong extends Boss {
     }
 
    @Override
-    public int injured(Player plAtt, int damage, boolean piercing, boolean isMobAttack) {
+    public long injured(Player plAtt, long damage, boolean piercing, boolean isMobAttack) {
  this.checkAnThan(plAtt);
         if (Util.isTrue(10, 100) && plAtt != null) {//tỉ lệ hụt của thiên sứ
             Util.isTrue(this.nPoint.tlNeDon, 100);
