+++ src/red/team/boss/list_boss/conuong/So18.java
@@ -61,7 +61,7 @@ public class So18 extends Boss {
         TaskService.gI().checkDoneTaskKillBoss(plKill, this);
     }
 @Override
-    public int injured(Player plAtt, int damage, boolean piercing, boolean isMobAttack) {
+    public long injured(Player plAtt, long damage, boolean piercing, boolean isMobAttack) {
  this.checkAnThan(plAtt);
         if (Util.isTrue(10, 100) && plAtt != null) {//tỉ lệ hụt của thiên sứ
             Util.isTrue(this.nPoint.tlNeDon, 100);
