+++ src/red/team/boss/list_boss/doanh_trai/TrungUyXanhLo.java
@@ -40,7 +40,7 @@ public class TrungUyXanhLo extends TrungUyTrang {
             ConstPlayer.TRAI_DAT, //gender
             new short[]{135, 136, 137, -1, -1, -1}, //outfit {head, body, leg, bag, aura, eff}
             500, //dame
-            new int[]{500}, //hp
+            new long[]{500}, //hp
             new int[]{1}, //map join
             new int[][]{
                 {Skill.DRAGON, 3, 1000},
@@ -71,7 +71,7 @@ public class TrungUyXanhLo extends TrungUyTrang {
     }
 
    @Override
-    public int injured(Player plAtt, int damage, boolean piercing, boolean isMobAttack) {
+    public long injured(Player plAtt, long damage, boolean piercing, boolean isMobAttack) {
  this.checkAnThan(plAtt);
         if (!this.isDie()) {
             if (!piercing && Util.isTrue(30, 100)) {
