+++ src/red/team/boss/list_boss/giangho/GIANGHO.java
@@ -13,11 +13,10 @@ import red.services.Service;
 import red.services.TaskService;
 import red.utils.Util;
 
-
 public class GIANGHO extends Boss {
 
     public GIANGHO() throws Exception {
-        super(BossID.GIANGHO, BossesData.SOI_GIANGHO, BossesData.ODO_GIANGHO, BossesData.XINBATO_GIANGHO, BossesData.CHACHA_GIANGHO, BossesData.PONPUT_GIANGHO,BossesData.CHANXU_GIANGHO,BossesData.TAUPAYPAY_GIANGHO,BossesData.YAMCHA_GIANGHO,BossesData.TAUPAYPAY_GIANGHO,BossesData.JACKYCHUN_GIANGHO,BossesData.THIENXINHANG_GIANGHO);
+        super(BossID.GIANGHO, BossesData.SOI_GIANGHO, BossesData.ODO_GIANGHO, BossesData.XINBATO_GIANGHO, BossesData.CHACHA_GIANGHO, BossesData.PONPUT_GIANGHO, BossesData.CHANXU_GIANGHO, BossesData.TAUPAYPAY_GIANGHO, BossesData.YAMCHA_GIANGHO, BossesData.TAUPAYPAY_GIANGHO, BossesData.JACKYCHUN_GIANGHO, BossesData.THIENXINHANG_GIANGHO);
     }
 
     @Override
@@ -26,36 +25,37 @@ public class GIANGHO extends Boss {
         if (this.currentLevel == 1) {
             return;
         }
-        int[] itemDos = new int[]{457,555,556,557,558,559,560,561,562,563,564,565,566,567};
-        int[] NRs = new int[]{15,16,16,16,16,16,16,16};
+        int[] itemDos = new int[]{457, 555, 556, 557, 558, 559, 560, 561, 562, 563, 564, 565, 566, 567};
+        int[] NRs = new int[]{15, 16, 16, 16, 16, 16, 16, 16};
         int randomDo = new Random().nextInt(itemDos.length);
         int randomNR = new Random().nextInt(NRs.length);
         if (Util.isTrue(10, 100)) {
-          
+
 //                if (Util.isTrue(2, 10)) {
 //                    Service.gI().dropItemMap(this.zone, Util.ratiItem(zone, 2000+plKill.gender, 1, this.location.x, this.location.y, plKill.id));
 //                }
 //            
             if (Util.isTrue(1, 3)) {
-                Service.gI().dropItemMap(this.zone, Util.ratiItem(zone, 344, 1, this.location.x+2, this.location.y, plKill.id));
+                Service.gI().dropItemMap(this.zone, Util.ratiItem(zone, 344, 1, this.location.x + 2, this.location.y, plKill.id));
             }
             Service.gI().dropItemMap(this.zone, Util.ratiItem(zone, itemDos[randomDo], 1, this.location.x, this.location.y, plKill.id));
-             } else {
+        } else {
             Service.gI().dropItemMap(this.zone, new ItemMap(zone, NRs[randomNR], 1, this.location.x, zone.map.yPhysicInTop(this.location.x, this.location.y - 24), plKill.id));
         }
         plKill.pointBoss += 0;
 
         TaskService.gI().checkDoneTaskKillBoss(plKill, this);
     }
-   @Override
-    public int injured(Player plAtt, int damage, boolean piercing, boolean isMobAttack) {
- this.checkAnThan(plAtt);
+
+    @Override
+    public long injured(Player plAtt, long damage, boolean piercing, boolean isMobAttack) {
+        this.checkAnThan(plAtt);
         if (!this.isDie()) {
             if (!piercing && Util.isTrue(this.nPoint.tlNeDon, 100)) {
                 this.chat("Xí hụt");
                 return 0;
             }
-           
+
             this.nPoint.subHP(damage);
             if (isDie()) {
                 this.setDie(plAtt);
@@ -66,6 +66,7 @@ public class GIANGHO extends Boss {
             return 0;
         }
     }
+
     @Override
     protected void notifyJoinMap() {
         if (this.currentLevel == 1) {
@@ -73,12 +74,13 @@ public class GIANGHO extends Boss {
         }
         super.notifyJoinMap();
     }
+
     @Override
     public void active() {
         super.active(); //To change body of generated methods, choose Tools | Templates.
-       if (Util.canDoWithTime(st, 600000)) {
-           this.changeStatus(BossStatus.LEAVE_MAP);
-       }
+        if (Util.canDoWithTime(st, 600000)) {
+            this.changeStatus(BossStatus.LEAVE_MAP);
+        }
     }
 
     @Override
