+++ src/red/team/boss/list_boss/nhanban/Namecnhanban.java
@@ -25,7 +25,7 @@ public class Namecnhanban extends Boss {
 
     @Override
     public void reward(Player plKill) {
-
+        
     }
 
     @Override
@@ -55,7 +55,7 @@ public class Namecnhanban extends Boss {
     }
 
     @Override
-    public int injured(Player plAtt, int damage, boolean piercing, boolean isMobAttack) {
+    public long injured(Player plAtt, long damage, boolean piercing, boolean isMobAttack) {
         this.checkAnThan(plAtt);
 
         if (!this.isDie()) {
