+++ src/red/team/boss/list_boss/tanjiro/Tanjiro.java
@@ -52,7 +52,7 @@ public class Tanjiro extends Boss {
         }
     }
    @Override
-    public int injured(Player plAtt, int damage, boolean piercing, boolean isMobAttack) {
+  public long injured(Player plAtt, long damage, boolean piercing, boolean isMobAttack) {
  this.checkAnThan(plAtt);
         if (!this.isDie()) {
             if (!piercing && Util.isTrue(this.nPoint.tlNeDon, 100)) {
